ai-chatbot/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
copier==8.3.0
daemon==1.2
dataclasses-json==0.6.1
decorator==5.1.1
docutils==0.20.1
dunamai==1.19.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.101.1
flet==0.10.3
flet-core==0.10.3
flet-fastapi==0.10.3
flet-runtime==0.10.3
frozenlist==1.4.0
funcy==2.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.311
langsmith==0.0.43
lockfile==0.12.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
pathspec==0.11.2
plumbum==1.8.2
prompt-toolkit==3.0.36
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3
Pygments==2.16.1
pypdf==3.16.4
pypng==0.20220715.0
python-daemon==3.0.1
python-dotenv==1.0.0
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.1
regex==2023.10.3
repath==0.9.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-extensions==4.8.0
typing-inspect==0.9.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websocket-client==1.6.4
websockets==11.0.3
yarl==1.9.2
